Shell: Do not raise ActiveDocumentChanging if same document gets activated again
There are problems with writing code that opens and closes windows based
on the active document if the ActiveDocumentChanging event occurs when the
same document is activated again. In my case I opened a new tool window
which would disappear right after ShowTool because it would raise
ActiveDocumentChanging with the same document again.

Signed-off-by: Axel Gembe <axel@gembe.net>