Inspector: Add SelectiveUndoEditorBase to disable undo selectively
This class is used for values that should only be updated after the user
has finished editing them. The view needs to call OnBeginEdit when the
user has started editing to capture the current value and call OnEndEdit
to commit the old and new value to the undo / redo manager.

This will be used in the next patch that makes the range editor not update
the undo / redo manager while the user is dragging the slider control.

Signed-off-by: Axel Gembe <axel@gembe.net>