Add German and Chinese (Simplified) localization to Gemini
This handles the inspector and output module as well as the Gemini core.
There were quite a few places where user visible strings were hardcoded
which have been changed to use a resource file. This also adds a
dependency on Gu.Wpf.Localization, which after careful evaluation seems to
be the best fit for Gemini because it is MIT licensed and under active
maintenance. Gu.Wpf.Localization is used to localize the xaml parts of
Gemini.

The settings were enhanced by adding a new combo box to select the active
language, which requires a restart to change. The default is to use the
current system language.

Signed-off-by: Axel Gembe <axel@gembe.net>