Undo Redo: Improve performance when undo / redo stack changes
I noticed while profiling my code that 99% of the CPU time to change the
undo redo stack is spent in Gemini instead of actually setting values.

First, I noticed that the view model gets refreshed twice for each single
change in the undo redo stack as the model is subscribed to a separate
event for each stack collection. Nothing easy that I could think of can be
done about this, but it is also not critical if we can optimize elsewhere.

Second, I noticed that the update of the history items in RefreshHistory
takes a long time due to data binding. I changed this by creating a new
list each time the view is refreshed and discarding the old list. After
the refresh, the HistoryItems property is notified so that the view can
update. This caused a significant increase in performance.

Third, I used the new batch notifications to suspend updates while a batch
is in progress and refresh the view after the batch has finished. This
improved my time for big undo / redo operations dramatically, although now
there is no more nice animating of the history view to see where your undo
is right now.

Signed-off-by: Axel Gembe <axel@gembe.net>