Shell: Fix duplicate ActiveDocumentChanged
The shell calls ActivateItem inside the ActiveLayoutItem setter which
causes ActiveDocumentChanged to be raised twice because ActivateItem calls
base.ActivateItem(item) which sets ActiveLayoutItem which calls
ActivateItem where the ReferenceEquals check fails to detect the duplicate
activation because ActiveItem is not updated yet.

This patch adds a guard in ActivateItem that prevents recursive calls.

Signed-off-by: Axel Gembe <axel@gembe.net>