Undo Redo: Add optional limit for number of undo entries
This patch adds the UndoCountLimit property to the UndoRedoManager which
is used to limit the maximum length of the stack. By default it is set to
null which means that no limit is enforced. If there is a value set, the
undo stack is always limited to this length but the redo stack is not
limited by this as it should not get longer than the undo stack anyway as
long as the limit is not changed.

Signed-off-by: Axel Gembe <axel@gembe.net>