Fix reset button being attached to the group object in groups
The reset button only worked for objects that are not in a collapsible
group. When an object was in a group clicking the reset button resulted
in invoking the Reset method on the group instead of the object.

The solution seems to be to refresh the Action.TargetWithoutContext to
the current binding.