Made SettingViewModel non-shared
When you open the Tools->Options and change a setting then click Cancel the user would assume that opening the Options window again would show the original setting however it currently shows the unsaved settings that the user selected before clicking Cancel. Changing the PartCreationPolicy to NonShared fixes this issue.