Fix PublishSingleFile cases
Fix PublishSingleFile under .netcoreapp3.1. Files are extracted to a temp folder, so in addition to the CWD, AppContext.BaseDirectory must be enumerated by a DirectoryCatalog

Fix PublishSingleFile under .net5.0-windows (and later, I imagine). Unlike .netcore, files are NOT extracted, unless you specify IncludeAllContentForSelfExtract=true in the .pubxml.

Add new DemoAppBootstrapper implementation to showcase how to support PublishSingleFile cases under .NET5. It is not used by default, requires user to change App.xaml.

Add .pubxml files for testing PublishSingleFile under .netcoreapp3.1 and .net5.0-windows.
Add vscode\launch.json for easy launch-and-debugging (with VS Code).

Under .NET5+, when PublishSingleFile is detected and there's no debugger, Gemini.AppBootstrapper.Config will present a message box with the message which also gets thrown as an exception. This is because when not under a debugger, the app crashes with no information (Gemini.Demo.App crash). When a debugger is present, only the exception is thrown.