Fix cases of readonly on [Import]'s
Both OpenDocumentResult and ShowToolResult imported IShell but their member variable was set to readonly, which resulted in a composition exception (member is not writable). They are no longer readonly.

Additionally. no code was actually making use of the Show.Document/Tool helpers, which is likely how this went uncaught, despite the README.md providing an example of it for the HomeViewModel. ViewHome and ViewToolboxCommandHandler now use Show class instead of importing IShell themselves.